// src/app/api/model/jogadorDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface JogadorDTO { 
    id?: number;
    eventoId?: number;
    timeId?: number;
    nome?: string;
    nivel?: JogadorDTO.NivelEnum;
    foto?: string;
}
export namespace JogadorDTO {
    export const NivelEnum = {
        Craque: 'CRAQUE',
        Mediano: 'MEDIANO',
        PernaDePau: 'PERNA_DE_PAU'
    } as const;
    export type NivelEnum = typeof NivelEnum[keyof typeof NivelEnum];
}